# internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.28.0

package db

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Pet struct {
	ID         pgtype.UUID        `json:"id"`
	Name       string             `json:"name"`
	Species    string             `json:"species"`
	Age        int32              `json:"age"`
	Status     string             `json:"status"`
	PhotoUrls  []string           `json:"photo_urls"`
	Tags       []string           `json:"tags"`
	CreatedAt  pgtype.Timestamptz `json:"created_at"`
	ModifiedAt pgtype.Timestamptz `json:"modified_at"`
	CreatedBy  string             `json:"created_by"`
	ModifiedBy string             `json:"modified_by"`
}